api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
playwright>=1.41.0
trafilatura>=1.6.0
readability-lxml>=0.8.1
html2text>=2024.2.26
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
orjson>=3.9.0
ollama>=0.1.6
